Factored out file related utilities.
