+++ software/util/copystaticdocsplusinsert.py
@@ -15,6 +15,8 @@ import re
 for path in [os.getcwd(),"software/SchemaTerms","software/SchemaExamples","software/util"]:
   sys.path.insert(1, path) #Pickup libs from local directories
 
+import fileutils
+
 if os.path.basename(os.getcwd()) != "schemaorg":
     print("\nScript should be run from within the 'schemaorg' directory! - Exiting\n")
     sys.exit(os.EX_DATAERR)
@@ -41,65 +43,11 @@ for f_path in glob.glob('./templates/static-doc-inserts/*.html'):
     indata = indata.replace('{{docsdir}}', "/docs")
     INSERTS[fn] = indata
 
-def mycopytree(src, dst, symlinks=False, ignore=None):
-    """Copy a full tree in the file-system.
-
-    This is basically a custom version of `shutils.copytree()`.
-
-    Parameters:
-        src (str): source path
-        dst (str): destination path
-        symlinks (bool): should symbolic links be followed.
-        ignore (callable): function that takes the set of filenames and returns those to be ignored.
-
-    """
-    #copes with already existing directories
-    names = os.listdir(src)
-    if ignore is not None:
-        ignored_names = ignore(src, names)
-    else:
-        ignored_names = frozenset()
-
-    if not os.path.isdir(dst):
-        os.makedirs(dst)
-    errors = []
-    for name in names:
-        if name in ignored_names:
-            continue
-        srcname = os.path.join(src, name)
-        dstname = os.path.join(dst, name)
-        try:
-            if symlinks and os.path.islink(srcname):
-                linkto = os.readlink(srcname)
-                os.symlink(linkto, dstname)
-            elif os.path.isdir(srcname):
-                mycopytree(srcname, dstname, symlinks, ignore)
-            else:
-                # Will raise a SpecialFileError for unsupported file types
-                copy2(srcname, dstname)
-        # catch the Error from the recursive copytree so that we can
-        # continue with other files
-        except Error as err:
-            errors.extend(err.args[0])
-        except EnvironmentError as why:
-            errors.append((srcname, dstname, str(why)))
-    try:
-        copystat(src, dst)
-    except OSError as why:
-        if WindowsError is not None and isinstance(why, WindowsError):
-            # Copying file access times may fail on Windows
-            pass
-        else:
-            errors.extend((src, dst, str(why)))
-    if errors:
-        raise Error(errors)
-
-
 SRCDIR = './docs'
 DESTDIR = './software/site/docs'
 
 def copydocs():
-    mycopytree(SRCDIR, DESTDIR)
+    fileutils.mycopytree(SRCDIR, DESTDIR)
 
 def htmlinserts():
     """Perform susbstitions on all HTML files in DESTDIR."""
